added checkout and assign user to cart items

+++ carts/views.py
@@ -1,6 +1,7 @@
 from django.shortcuts import render, redirect
 from store.models import Product, Variation
 from carts.models import Cart, CartItem
+from django.contrib.auth.decorators import login_required
 # Create your views here.
 
 
@@ -12,6 +13,10 @@ def _get_cart_id(request):
 
 
 def add_cart(request, product_id):
+    if request.user.is_authenticated:
+        user = request.user
+    else:
+        user = None
     product = Product.objects.get(id=product_id)
 
     new_variation = []
@@ -29,35 +34,47 @@ def add_cart(request, product_id):
     except Cart.DoesNotExist:
         cart = Cart.objects.create(cart_id=_get_cart_id(request))
 
-    cartItems = CartItem.objects.filter(product=product, cart=cart)
+    if request.user.is_authenticated:
+        cart_items = CartItem.objects.filter(product=product, user=user)
+    else:
+        cart_items = CartItem.objects.filter(product=product, cart=cart)
 
     existing_variation = []
     cartItem_id = []
-    for cartItem in cartItems:
+    for cartItem in cart_items:
         # this will add multiple Variation objects of particular cartItem
         existing_variation.append(list(cartItem.variations.all()))
         cartItem_id.append(cartItem.id)
-
-    if CartItem.objects.filter(product=product, cart=cart).exists() and new_variation in existing_variation:
+    print("Existing Var; ",existing_variation)
+    print("new variations: ",new_variation)
+    if (CartItem.objects.filter(product=product, cart=cart).exists() or CartItem.objects.filter(product=product, user=user).exists()) and new_variation in existing_variation:
         index = existing_variation.index(new_variation)
-        cartItem = CartItem.objects.get(product=product, cart=cart, id=cartItem_id[index])
+        if user is not None:
+            cartItem = CartItem.objects.get(product=product,user=user,id=cartItem_id[index])
+        else:
+            cartItem = CartItem.objects.get(product=product,id=cartItem_id[index])
         cartItem.quantity += 1
     else:
-        cartItem = CartItem.objects.create(cart=cart, product=product, quantity=1)
+        cartItem = CartItem.objects.create(user=user,cart=cart, product=product, quantity=1)
         cartItem.variations.add(*new_variation)
 
     cartItem.save()
     return redirect("cart")
 
 
-def cart(request, cart_items=None):
+
+#Cart Calculations
+def cart_calculation(request):
     try:
         tax = 0
         grand_total = 0
         sub_total = 0
-        cart = Cart.objects.get(cart_id=_get_cart_id(request))
-        cart_items = cart.cart_products.all()
 
+        if request.user.is_authenticated:
+            cart_items = CartItem.objects.filter(user=request.user)
+        else:
+            cart = Cart.objects.get(cart_id=_get_cart_id(request))
+            cart_items = CartItem.objects.filter(cart=cart)
         # calculate total price of products with quantity
         for item in cart_items:
             item.total_price = item.product.price * item.quantity
@@ -65,28 +82,38 @@ def cart(request, cart_items=None):
 
         tax = (grand_total * 20) / 100
         sub_total = grand_total + tax
-    except Cart.DoesNotExist:
-        pass  # ignore
-    context = {
+        context = {
         "cart_items": cart_items,
         "grand_total": grand_total,
         "tax": tax,
         "sub_total": sub_total,
     }
+    except Cart.DoesNotExist:
+        context = {}
+
+    return context
+
+def cart(request):
+    context = cart_calculation(request)
     return render(request, "store/cart.html", context)
 
 
 def decrement_item(request, cart_item):
     item = CartItem.objects.get(id=cart_item)
-
     if request.method == "POST":
-        item.delete()
+        if request.user.is_authenticated and item.user == request.user:
+            item.delete()
+        else:
+            if item.user is None and item.cart.cart_id == _get_cart_id(request):
+                item.delete()
+                
         return redirect("cart")
     else:
-        item.quantity -= 1
-        item.save()
-        if item.quantity == 0:
-            item.delete()
+        if (request.user.is_authenticated and item.user == request.user) or (item.user is None and item.cart.cart_id == _get_cart_id(request)):
+            item.quantity -= 1
+            item.save()
+            if item.quantity == 0:
+                item.delete()
         return redirect('cart')
 
 
@@ -94,3 +121,10 @@ def decrement_item(request, cart_item):
 #     item = CartItem.objects.get(id=cart_item)
 #     item.delete()
 #     return redirect("cart")
+
+
+#checkout page
+@login_required
+def checkout(request):
+    context = cart_calculation(request)
+    return render(request,"store/checkout.html",context)
